requirements.txt
httpx==0.28.1
pydantic==2.11.9
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.178.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
aiofiles==24.1.0
fastapi==0.116.1
uvicorn[standard]==0.35.0
starlette==0.47.3
anyio==4.10.0
# LangChain and RAG dependencies
langchain==0.3.27
langchain-core==0.3.78
langchain-experimental==0.3.4
langchain-text-splitters==0.3.11
langchain-community==0.3.29
langchain-chroma==0.2.6
chromadb==1.1.0
ragas==0.3.6
# Document dependecies
pypdf==6.1.0
unstructured==0.18.15
Markdown==3.9
# Preprocessing dependencies
markitdown==0.1.3
datasets==4.1.1
# Google AI dependencies
langchain-google-genai==2.0.10
google-generativeai==0.8.5
python-dotenv==1.1.1
# Reranking dependencies
cohere==5.18.0
langchain-cohere==0.4.6
sentence-transformers==5.1.1
# Query rewriting dependencies
langchain-huggingface==0.3.1
scikit-learn==1.7.2
ipywidgets==8.1.7
ipykernel==6.30.1
faiss-cpu==1.12.0
pydantic==2.11.9
